changed int to float in lux

+++ main.cpp
@@ -43,7 +43,7 @@ int main()
     std::cout << "config ADC in CHIP" << std::bitset<16>(ads1015.getConfig()) << std::endl;
     std::cout << "Written config ADC " << std::bitset<16>(config.reg) << std::endl;
 
-    int lux = ltr559.getLux();
+    float lux = ltr559.getLux();
     float voltage = ads1015.readADC();
 
     // initializing BME280
